requirements.txt
beir==2.1.0
numpy<2
pandas
python-dotenv